chore: finish renaming SheetSidebar to SheetContentSidebar
